Extract ActivityCardLoader class
