+++ src/main/java/seedu/taskman/ui/activitycard/ActivityCard.java
@@ -40,36 +40,6 @@ public class ActivityCard extends UiPart {
 
     }
 
-    /**
-     * Constructs the appropriate ActivityCard given the panelType
-     *
-     * @param activity       to be represented
-     * @param panelType      of the Panel
-     * @param displayedIndex of the activity in the panel
-     * @return the ActivityCard corresponding to the panelType
-     */
-    public static ActivityCard load(Activity activity, Activity.PanelType panelType, int displayedIndex) {
-        ActivityCard card = new ActivityCard();
-
-        switch (panelType) {
-            case DEADLINE: {
-                card = new DeadlineActivityCard();
-                break;
-            }
-            case SCHEDULE: {
-                card = new ScheduleActivityCard();
-                break;
-            }
-            case FLOATING: {
-                card = new FloatingActivityCard();
-                break;
-            }
-        }
-        card.activity = activity;
-        card.displayedIndex = displayedIndex;
-        return UiPartLoader.loadUiPart(card);
-    }
-
     @Override
     public void setNode(Node node) {
         cardPane = (GridPane) node;
